seamgui/webapp/requirements.txt
Flask
wtforms
flask-bootstrap
flask-appconfig
flask-wtf
flask-mail
pyyaml
